Implement delete account functionality in userService; add confirmation prompt, authentication check, and error handling
