+++ frontend/app/service/userService.ts
@@ -1,8 +1,7 @@
 import { authService } from "./authService";
 import type { User } from "~/model/user";
 
-export const 
-userService = {
+export const userService = {
     async createUser(user: User) {
         const jwtToken = await authService.getJwtToken();
         const response = await fetch("/api/users", {
@@ -46,6 +45,48 @@ userService = {
             throw new Error("Failed to update user");
         }
         return response.json();
+    },
+
+    async deleteAccount(): Promise<void> {
+        const confirmation = window.confirm("Are you sure you want to delete your account? This action cannot be undone.");
+        if (!confirmation) return;
+
+        try {
+            const isAuthenticated = await authService.isAuthenticated();
+            if (!isAuthenticated) {
+                alert("You are not authenticated. Please log in again.");
+                return;
+            }
+
+            const idToken = await authService.getJwtToken();
+            const userId = (await authService.getCurrentUser())?.uid;
+            if (!userId) {
+                alert("Unable to retrieve user ID. Please try again.");
+                return;
+            }
+
+            const response = await fetch(`/api/users/${userId}`, {
+                method: "DELETE",
+                headers: {
+                    "Content-Type": "application/json",
+                    "Authorization": `Bearer ${idToken}`,
+                },
+            });
+
+            if (response.status === 204) {
+                alert("Account deleted successfully.");
+                await authService.logout();
+                window.location.href = "/login";
+            } else {
+                const errorData = await response.json();
+                console.error("Error:", errorData);
+                alert(`Error deleting account: ${errorData.error}`);
+            }
+        } catch (error) {
+            console.error("Error:", error);
+            alert("An error occurred while deleting the account. Please try again later.");
+        }
     }
-}
+};
+export default userService;
